chore: move generate translation file logic to the cli
