+++ packages/pieces/community/framework/translation-keys.json
@@ -1,0 +1,22 @@
+[
+    "displayName",
+    "description",
+    "auth.username.displayName",
+    "auth.username.description",
+    "auth.password.displayName",
+    "auth.password.description",
+    "auth.props.*.displayName",
+    "auth.props.*.description",
+    "auth.props.*.options.options.*.label",
+    "auth.description",
+    "actions.*.displayName",
+    "actions.*.description",
+    "actions.*.props.*.displayName",
+    "actions.*.props.*.description",
+    "actions.*.props.*.options.options.*.label",
+    "triggers.*.displayName",
+    "triggers.*.description",
+    "triggers.*.props.*.displayName",
+    "triggers.*.props.*.description",
+    "triggers.*.props.*.options.options.*.label"
+]
